Split settings utils further, want to be able to call is_running_tests ASAP, but being inside a module that implicitly relied on being instantiated post-django settings configuration
