+++ phac_aspc/django/settings/utils/__init__.py
@@ -1,5 +1,6 @@
 # flake8: noqa
 """Settings utils module"""
-from .configuration_utils import *
+from .configuration_verification_utils import *
+from .configure_settings_for_tests import *
 from .env_utils import *
-from .test_utils import *
+from .is_running_tests import *
